Upgrade to Spring Boot 2.0 RC1

+++ genie-web/src/main/java/com/netflix/genie/web/jpa/services/JpaJobSearchServiceImpl.java
@@ -194,7 +194,7 @@ public class JpaJobSearchServiceImpl implements JobSearchService {
 
             final List<JobSearchResult> results = this.entityManager
                 .createQuery(contentQuery)
-                .setFirstResult(page.getOffset())
+                .setFirstResult(((Long) page.getOffset()).intValue())
                 .setMaxResults(page.getPageSize())
                 .getResultList();
